♻️ Refact: Error 관련 파일 폴더링
